requirements.txt
aiodns==3.1.1
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
beautifulsoup4==4.12.2
attrs==23.1.0
cachetools==5.3.2
ccxt==4.1.63
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dateparser==1.2.0
distlib==0.3.7
distro==1.8.0
filelock==3.13.1
fonttools==4.45.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
openai==1.3.6
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
pycares==4.4.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
pyproject-api==1.6.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tox==4.11.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.1.0
virtualenv==20.24.7
websockets==12.0
yarl==1.9.2
